requirements.txt
tensorflow
tensorflow-datasets
matplotlib
pandas
numpy
scikit-learn
jupyter
pillow
streamlit
seaborn
